Placed the log in logic into a seperate function inside service directory/file Fixes #21
